requirements.txt
googletrans==4.0.0-rc1
nltk
robotframework
selenium
robotframework-seleniumlibrary
robotframework-pabot
bs4
requests
